move get_limited_unpacker to helpers
also: move some constants to borg.constants
